refactor array scoped predicate operators
