+++ crates/rulemorph/src/transform/operators/array/scoped.rs
@@ -1,5 +1,7 @@
 use super::*;
 
+pub(super) mod predicate;
+
 pub(in crate::transform::operators) fn eval_array_map(
     args: &[Expr],
     injected: Option<&EvalValue>,
@@ -34,46 +36,6 @@ pub(in crate::transform::operators) fn eval_array_map(
         let item_locals = locals_with_item(locals, EvalItem { value: item, index });
         let value = eval_expr_or_null(expr, record, context, out, &expr_path, Some(&item_locals))?;
         results.push(value);
-    }
-
-    Ok(EvalValue::Value(JsonValue::Array(results)))
-}
-
-pub(in crate::transform::operators) fn eval_array_filter(
-    args: &[Expr],
-    injected: Option<&EvalValue>,
-    record: &JsonValue,
-    context: Option<&JsonValue>,
-    out: &JsonValue,
-    base_path: &str,
-    locals: Option<&EvalLocals<'_>>,
-) -> Result<EvalValue, TransformError> {
-    let total_len = args_len(args, injected);
-    if total_len != 2 {
-        return Err(TransformError::new(
-            TransformErrorKind::ExprError,
-            "expr.args must contain exactly two items",
-        )
-        .with_path(format!("{}.args", base_path)));
-    }
-
-    let array = eval_array_arg(0, args, injected, record, context, out, base_path, locals)?;
-    let expr = arg_expr_at(1, args, injected).ok_or_else(|| {
-        TransformError::new(
-            TransformErrorKind::ExprError,
-            "expr.args index is out of bounds",
-        )
-        .with_path(format!("{}.args[1]", base_path))
-    })?;
-    let expr_index = if injected.is_some() { 0 } else { 1 };
-    let expr_path = format!("{}.args[{}]", base_path, expr_index);
-
-    let mut results = Vec::new();
-    for (index, item) in array.iter().enumerate() {
-        let item_locals = locals_with_item(locals, EvalItem { value: item, index });
-        if eval_predicate_expr(expr, record, context, out, &expr_path, Some(&item_locals))? {
-            results.push(item.clone());
-        }
     }
 
     Ok(EvalValue::Value(JsonValue::Array(results)))
@@ -202,52 +164,6 @@ pub(in crate::transform::operators) fn eval_array_key_by(
     }
 
     Ok(EvalValue::Value(JsonValue::Object(results)))
-}
-
-pub(in crate::transform::operators) fn eval_array_partition(
-    args: &[Expr],
-    injected: Option<&EvalValue>,
-    record: &JsonValue,
-    context: Option<&JsonValue>,
-    out: &JsonValue,
-    base_path: &str,
-    locals: Option<&EvalLocals<'_>>,
-) -> Result<EvalValue, TransformError> {
-    let total_len = args_len(args, injected);
-    if total_len != 2 {
-        return Err(TransformError::new(
-            TransformErrorKind::ExprError,
-            "expr.args must contain exactly two items",
-        )
-        .with_path(format!("{}.args", base_path)));
-    }
-
-    let array = eval_array_arg(0, args, injected, record, context, out, base_path, locals)?;
-    let expr = arg_expr_at(1, args, injected).ok_or_else(|| {
-        TransformError::new(
-            TransformErrorKind::ExprError,
-            "expr.args index is out of bounds",
-        )
-        .with_path(format!("{}.args[1]", base_path))
-    })?;
-    let expr_index = if injected.is_some() { 0 } else { 1 };
-    let expr_path = format!("{}.args[{}]", base_path, expr_index);
-
-    let mut matched = Vec::new();
-    let mut unmatched = Vec::new();
-    for (index, item) in array.iter().enumerate() {
-        let item_locals = locals_with_item(locals, EvalItem { value: item, index });
-        if eval_predicate_expr(expr, record, context, out, &expr_path, Some(&item_locals))? {
-            matched.push(item.clone());
-        } else {
-            unmatched.push(item.clone());
-        }
-    }
-
-    Ok(EvalValue::Value(JsonValue::Array(vec![
-        JsonValue::Array(matched),
-        JsonValue::Array(unmatched),
-    ])))
 }
 
 pub(in crate::transform::operators) fn eval_array_distinct_by(
@@ -387,84 +303,6 @@ pub(in crate::transform::operators) fn eval_array_sort_by(
 
     let results = items.into_iter().map(|item| item.value).collect::<Vec<_>>();
     Ok(EvalValue::Value(JsonValue::Array(results)))
-}
-
-pub(in crate::transform::operators) fn eval_array_find(
-    args: &[Expr],
-    injected: Option<&EvalValue>,
-    record: &JsonValue,
-    context: Option<&JsonValue>,
-    out: &JsonValue,
-    base_path: &str,
-    locals: Option<&EvalLocals<'_>>,
-) -> Result<EvalValue, TransformError> {
-    let total_len = args_len(args, injected);
-    if total_len != 2 {
-        return Err(TransformError::new(
-            TransformErrorKind::ExprError,
-            "expr.args must contain exactly two items",
-        )
-        .with_path(format!("{}.args", base_path)));
-    }
-
-    let array = eval_array_arg(0, args, injected, record, context, out, base_path, locals)?;
-    let expr = arg_expr_at(1, args, injected).ok_or_else(|| {
-        TransformError::new(
-            TransformErrorKind::ExprError,
-            "expr.args index is out of bounds",
-        )
-        .with_path(format!("{}.args[1]", base_path))
-    })?;
-    let expr_index = if injected.is_some() { 0 } else { 1 };
-    let expr_path = format!("{}.args[{}]", base_path, expr_index);
-
-    for (index, item) in array.iter().enumerate() {
-        let item_locals = locals_with_item(locals, EvalItem { value: item, index });
-        if eval_predicate_expr(expr, record, context, out, &expr_path, Some(&item_locals))? {
-            return Ok(EvalValue::Value(item.clone()));
-        }
-    }
-
-    Ok(EvalValue::Value(JsonValue::Null))
-}
-
-pub(in crate::transform::operators) fn eval_array_find_index(
-    args: &[Expr],
-    injected: Option<&EvalValue>,
-    record: &JsonValue,
-    context: Option<&JsonValue>,
-    out: &JsonValue,
-    base_path: &str,
-    locals: Option<&EvalLocals<'_>>,
-) -> Result<EvalValue, TransformError> {
-    let total_len = args_len(args, injected);
-    if total_len != 2 {
-        return Err(TransformError::new(
-            TransformErrorKind::ExprError,
-            "expr.args must contain exactly two items",
-        )
-        .with_path(format!("{}.args", base_path)));
-    }
-
-    let array = eval_array_arg(0, args, injected, record, context, out, base_path, locals)?;
-    let expr = arg_expr_at(1, args, injected).ok_or_else(|| {
-        TransformError::new(
-            TransformErrorKind::ExprError,
-            "expr.args index is out of bounds",
-        )
-        .with_path(format!("{}.args[1]", base_path))
-    })?;
-    let expr_index = if injected.is_some() { 0 } else { 1 };
-    let expr_path = format!("{}.args[{}]", base_path, expr_index);
-
-    for (index, item) in array.iter().enumerate() {
-        let item_locals = locals_with_item(locals, EvalItem { value: item, index });
-        if eval_predicate_expr(expr, record, context, out, &expr_path, Some(&item_locals))? {
-            return Ok(EvalValue::Value(JsonValue::Number((index as i64).into())));
-        }
-    }
-
-    Ok(EvalValue::Value(JsonValue::Number((-1).into())))
 }
 
 pub(in crate::transform::operators) fn eval_array_reduce(
